# ms-order/src/main/java/br/com/pb/msorder/domain/dto/AddressRequest.java
package br.com.pb.msorder.domain.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class AddressRequest {

    private String cep;
    private String numero;
    private String logradouro;
    private String bairro;
}